% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{Array}
\alias{Array}
\title{Array}
\usage{
Array(a, type = "f32")
}
\arguments{
\item{a}{array which columns have same types and same dimensions.}

\item{type}{Desired khiva Array type. By default: f32}
}
\value{
A KHIVA array created from the data stored in the array.
}
\description{
Creates a khiva Array from a R array.
}
