% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{C3}
\alias{C3}
\title{C3}
\usage{
C3(arr, lag)
}
\arguments{
\item{arr}{KHIVA array with the time series.}

\item{lag}{The lag.}
}
\value{
The non-linearity value for the given time series.
}
\description{
Calculates the Schreiber, T. and Schmitz, A. (1997) measure of non-linearity
for the given time series.
}
