% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{CidCe}
\alias{CidCe}
\title{CicCe}
\usage{
CidCe(arr, z.normalize)
}
\arguments{
\item{arr}{KHIVA array with the time series.}

\item{z.normalize}{Controls whether the time series should be z-normalized or not.}
}
\value{
The complexity value for the given time series.
}
\description{
Calculates an estimate for the time series complexity defined by
Batista, Gustavo EAPA, et al (2014). (A more complex time series has more peaks,
valleys, etc.)
}
