% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{CrossCovariance}
\alias{CrossCovariance}
\title{CrossCovariance}
\usage{
CrossCovariance(arr.xss, arr.yss, unbiased)
}
\arguments{
\item{arr.xss}{KHIVA array with the time series.}

\item{arr.yss}{KHIVA array with the time series.}

\item{unbiased}{Determines whether it divides by n - lag (if true) or n (if false).}
}
\value{
The cross-covariance value for the given time series.
}
\description{
Calculates the cross-covariance of the given time series.
}
