% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{methods}
\name{\%\%,Array,Array-method}
\alias{\%\%,Array,Array-method}
\title{Performs the modulo operation.}
\usage{
\S4method{\%\%}{Array,Array}(e1, e2)
}
\description{
Performs the modulo operation.
}
