% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Kurtosis}
\alias{Kurtosis}
\title{Kurtosis}
\usage{
Kurtosis(arr)
}
\arguments{
\item{arr}{KHIVA array with the time series.}
}
\value{
The kurtosis of each arr.
}
\description{
Returns the kurtosis of arr (calculated with the adjusted Fisher-Pearson
standardized moment coefficient G2).
}
