% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{LinearTrend}
\alias{LinearTrend}
\title{LinearTrend}
\usage{
LinearTrend(arr)
}
\arguments{
\item{arr}{KHIVA array with the time series.}
}
\value{
List of KHIVA Array with:
{
    pvalue: The pvalues for all time series.
    rvalue: The rvalues for all time series.
    intercept: The intercept values for all time series.
    slope: The slope for all time series.
    stdrr: The stderr values for all time series.
}
}
\description{
Calculates a linear least-squares regression for the values of the time series versus the sequence from 0 to
length of the time series minus one.
}
