% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{LongestStrikeBelowMean}
\alias{LongestStrikeBelowMean}
\title{LongestStrikeBelowMean}
\usage{
LongestStrikeBelowMean(arr)
}
\arguments{
\item{arr}{KHIVA array with the time series.}
}
\value{
The length of the longest consecutive subsequence in the input time series that is below the mean.
}
\description{
Calculates the length of the longest consecutive subsequence in arr that is below the mean of arr.
}
