% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{MeanSecondDerivativeCentral}
\alias{MeanSecondDerivativeCentral}
\title{MeanSecondDerivativeCentral}
\usage{
MeanSecondDerivativeCentral(arr)
}
\arguments{
\item{arr}{KHIVA array with the time series.}
}
\value{
The mean value of a central approximation of
the second derivative for each time series.
}
\description{
Calculates mean value of a central approximation of the second derivative for each time series in arr.
}
