% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{MomentStatistics}
\alias{MomentStatistics}
\title{MomentStatistics}
\usage{
MomentStatistics(tss, k)
}
\arguments{
\item{tss}{Expects an input array whose dimension zero is the length of the time series (all the same) and
dimension one indicates the number of time series.}

\item{k}{The specific moment to be calculated.}
}
\value{
The kth moment of the given time series.
}
\description{
Returns the kth moment of the given time series.
}
