% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality.R
\name{Pip}
\alias{Pip}
\title{Pip}
\usage{
Pip(arr, numberIPs)
}
\arguments{
\item{arr}{KHIVA Array whose dimension zero is the length of the time series.}

\item{numberIPs}{The number of points to be returned.}
}
\value{
Array with the most Perceptually Important numberIPs.
}
\description{
Calculates the number of Perceptually Important Points (PIP) in the time series.
}
\details{
[1] Fu TC, Chung FL, Luk R, and Ng CM. Representing financial time series based on data point importance.
Engineering Applications of Artificial Intelligence, 21(2):277-300, 2008.
}
