% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{RatioValueNumberToTimeSeriesLength}
\alias{RatioValueNumberToTimeSeriesLength}
\title{RatioValueNumberToTimeSeriesLength}
\usage{
RatioValueNumberToTimeSeriesLength(arr)
}
\arguments{
\item{arr}{KHIVA array with the time series.}
}
\value{
KHIVA array with the ratio of unique values with respect to the total number of values.
}
\description{
Calculates a factor which is 1 if all values in the time series occur only once, and below one if this is
not the case. In principle, it just returns:
}
\details{
\deqn{
\frac{\textit{number\_unique\_values}}{\textit{number\_values}}
}
}
