% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{SkewnessStatistics}
\alias{SkewnessStatistics}
\title{SkewnessStatistics}
\usage{
SkewnessStatistics(tss)
}
\arguments{
\item{tss}{Expects an input array whose dimension zero is the length of the time series (all the same) and
dimension one indicates the number of time series.}
}
\value{
Array containing the skewness of each time series in tss.
}
\description{
Calculates the sample skewness of tss (calculated with the adjusted Fisher-Pearson standardized moment
coefficient G1).
}
