% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{StandardDeviation}
\alias{StandardDeviation}
\title{StandardDeviation}
\usage{
StandardDeviation(arr)
}
\arguments{
\item{arr}{KHIVA array with the time series.}
}
\value{
The standard deviation of each time series within arr.
}
\description{
Calculates the standard deviation of each time series within arr.
}
