% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{StompSelfJoin}
\alias{StompSelfJoin}
\title{StompSelfJoin}
\usage{
StompSelfJoin(t, m)
}
\arguments{
\item{t}{KHIVA Array the time series.}

\item{m}{KHIVA Array the time series.}
}
\value{
List of KHIVA Arrays with the matrix profile and the index profile
}
\description{
STOMP algorithm to calculate the matrix profile between 't' and itself using a subsequence length
of 'm'. This method filters the trivial matches.
}
